requirements.txt
flask
flask-socketio
pypdf
numpy
faiss-cpu
requests
eventlet
